feat(skeleton): move use-skeleton to skeleton dir
