+++ packages/skeleton/src/hooks/use-skeleton/index.ts
@@ -1,0 +1,1 @@
+export { useSkeleton } from './use-skeleton';
