+++ packages/skeleton/src/hooks/use-skeleton/use-skeleton.module.css
@@ -1,4 +1,4 @@
-@import '../../../vars/src/index.css';
+@import '../../../../vars/src/index.css';
 
 .skeletonText.skeletonText {
     border-radius: var(--border-radius-pill);
