+++ packages/skeleton/src/index.ts
@@ -1,1 +1,4 @@
 export * from './Component';
+
+export { useSkeleton } from './hooks/use-skeleton';
+export type { TextSkeletonProps } from './types/text-skeleton-props';
