+++ packages/skeleton/src/types/text-skeleton-props.ts
@@ -1,0 +1,23 @@
+type WidthUnit = number | string;
+
+export type TextSkeletonProps = {
+    /**
+     *  Кол-во строк текста
+     */
+    rows?: number;
+
+    /**
+     * Ширина строки
+     */
+    width?: WidthUnit | WidthUnit[];
+
+    /**
+     * Выравнивание элементов по горизонтали
+     */
+    align?: 'left' | 'center' | 'right';
+
+    /**
+     * Класс для обертки скелетона
+     */
+    wrapperClassName?: string;
+};
