+++ packages/typography/src/text/component.tsx
@@ -2,9 +2,10 @@ import React, { forwardRef, HTMLAttributes } from 'react';
 import mergeRefs from 'react-merge-refs';
 import cn from 'classnames';
 
+import { type TextSkeletonProps, useSkeleton } from '@alfalab/core-components-skeleton';
+
 import { Color } from '../colors';
-import { useSkeleton } from '../hooks';
-import { TextElementType, TextSkeletonProps } from '../types';
+import { TextElementType } from '../types';
 
 import colors from '../colors.module.css';
 import styles from './index.module.css';
